Refactor common components into common/components
Refactor common components from features into `common/components`.

* **InfoModal Component:**
  - Add `InfoModal.jsx` to `src/common/components`.
  - Remove `InfoModal` implementation from `DashboardHeader.jsx`.
  - Import and use `InfoModal` in `DashboardHeader.jsx`.

* **TabSwitchBar Component:**
  - Add `TabSwitchBar.jsx` to `src/common/components`.
  - Remove tab switch bar implementation from `TopNavigation.jsx`.
  - Import and use `TabSwitchBar` in `TopNavigation.jsx`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/lion-agi/lion-studio?shareId=XXXX-XXXX-XXXX-XXXX).
